# Search file blocks rather than individual lines for faster results