<!--
Many thanks to contributing to MultiQC!
Please fill in the appropriate checklist below (delete whatever is not relevant).
-->

- [x] This comment contains a description of changes (with reason)

Searching file content is the most costly part of file searching for MultiQC. Currently all lines are searched individually and all lines are cached. 

However, this approach is problematic from a performance standpoint because of the overhead involved. Python iteration is calling the `__next__` function under the hood, the `content in line` calls `__contains__` under the hood. Python function calls are relatively expensive and this cost is payed for each line. Furthermore, string searching algorithms create a search structure first, and then iterate over the string using the structure to find matches. This preparation is done over and over again for each line. 

So rather than reading individual lines, this PR reads blocks of 4 KB from the file. To make sure the per-line-content matching is not broken, we use python's count and rfind method to ensure a block only contains full lines. 

The blocks are cached the same way as on the line iterator. 

This method may perform worse in the following cases:
- The content is right at the beginning of the file. Hence 4kb - the line size was processed too much.
- The line sizes are regularly bigger than 4kb. 

In the first case, this is not too problematic because said 4kb will be in the CPU L1 cache, and executing the count and rfind methods will be very fast. The second case is simply not likely, and optimization happens for the common case, not the edge case. 

EDIT: I forgot to mention that IO-wise, while more data is read at once, this does not make a change. Python buffers callls under the hood and already reads io.DEFAULT_BUFFER_SIZE of data from the file. This is 8 kb of raw data. Since we only read 4096 utf-8 characters, most of which will be one byte, there should be only one call to the raw file read.

The tradeoff is slightly more complex code but better runtimes. 

Before:
```
|           multiqc | Run took 24.08 seconds
|           multiqc |  - 6.65s: Searching files
|           multiqc |  - 7.96s: Running modules
|           multiqc |  - 7.25s: Compressing report data
```

After:
```
|           multiqc | Run took 22.49 seconds
|           multiqc |  - 5.47s: Searching files
|           multiqc |  - 7.98s: Running modules
|           multiqc |  - 6.82s: Compressing report data
```