@rhpvorderman 2024-04-29
@vladsavelyev Here it is! This is the last PR I have in the pipeline. I do plan to take a look at the MultiQC memory consumption. I have some ideas, but I have to see if they are applicable.