@vladsavelyev 2024-04-29
That's really awesome, I love the approach. And thanks for anticipating the edge cases.

In your benchmarking, I wonder why the compression is waster?