@rhpvorderman 2024-04-29
> In your benchmarking, I wonder why the compression is faster?

Run to run variance. The LZcompression does a lot of memory allocations (due to creating a massive dict). These are variably expensive. 
Should be solved with the gzip refactor.