@vladsavelyev 2024-04-29
Was wondering about the file search time - so we shouldn't expect similar run to run variance for it?

> Should be solved with the gzip refactor

Just merged it!