@vladsavelyev 2024-04-29
While testing, realized that https://github.com/MultiQC/test-data is not a good source of data to benchmark, as it only contains small toy examples.

I'm now thinking of creating a separate test bundle with some large real-life files to stress-test and benchmark MultiQC consistently. If you happen to have any such files handy, could you please share either in this issue, or to vladislav.sav@gmai.com?

(A bit unrelated, but we are also looking into improving the report performance for large datasets (https://github.com/MultiQC/MultiQC/issues/2280), so any huge examples would be great to have to stress-test that as well)