@vladsavelyev 2024-04-29
Tried on a larger dataset - the difference is really impressive!

Before:

```
25.55s: Searching files
```

After:

```
13.65s: Searching files
```