@rhpvorderman 2024-04-29
Wow!
What kind of larger dataset do you have available? I would love to use that to reduce the multiqc memory usage. Can you share it by any chance?