@ewels 2024-04-29
> The line sizes are regularly bigger than 4kb.

This isn't impossible - some files generate minified JSON outputs for example, where the entire file is a single line.

But the search should still work in these cases, right? As long as it doesn't completely break I'm happy. There are not many of them.

You'll probably ask me which tools do this and I can't remember off the top of my head sorry.