@rhpvorderman 2024-04-30
> But the search should still work in these cases, right? As long as it doesn't completely break I'm happy. There are not many of them.

Yes it will work just fine. It will only be suboptimal in a performance sense compared to letting python find the lines by using __next__ or readline. But for any case where the lines are less than 4096 bytes, the performance will be much better. So it is a tradeoff.